+++ lib/Applications.ps1
@@ -187,6 +187,7 @@ function app_status($app, $global) {
         $status.latest_version = $manifest.version
     }
 
+    $status.manifest = $manifest
     $status.outdated = $false
     if ($status.version -and $status.latest_version) {
         $status.outdated = (Compare-Version -ReferenceVersion $status.version -DifferenceVersion $status.latest_version) -ne 0